Update app.js
Replaced most of the sql calls with mongo. This should be a good starting point. Now we need to check all the functionlity provided by the end points.
Please search for "Anshul" in the app.js file to see which queries are remaning to be replaced by mongo. We should use this app.js for future.